Changes - package name changed - utils class removed and code moved to inst class - bug fix in servlet 6 httpsession inst
