--- instrumentation-security/low-severity-hooks/src/main/java/com/nr/agent/security/random/CryptoUtils.java
@@ -1,46 +1,0 @@
-package com.nr.agent.security.random;
-
-import com.newrelic.api.agent.security.NewRelicSecurity;
-import com.newrelic.api.agent.security.schema.AbstractOperation;
-import com.newrelic.api.agent.security.schema.SecurityMetaData;
-import com.newrelic.api.agent.security.schema.VulnerabilityCaseType;
-import com.newrelic.api.agent.security.schema.exceptions.NewRelicSecurityException;
-import com.newrelic.api.agent.security.schema.operation.HashCryptoOperation;
-
-public class CryptoUtils {
-    public static AbstractOperation preprocessSecurityHook(String algorithm, String provider, String className, String methodName, String category) {
-        try {
-            SecurityMetaData securityMetaData = NewRelicSecurity.getAgent().getSecurityMetaData();
-            if (!NewRelicSecurity.isHookProcessingActive() || securityMetaData.getRequest().isEmpty()
-            ) {
-                return null;
-            }
-
-            HashCryptoOperation operation = new HashCryptoOperation(algorithm, className, methodName, VulnerabilityCaseType.CRYPTO);
-            operation.setProvider(provider);
-            operation.setLowSeverityHook(true);
-            operation.setEventCategory(category);
-
-            NewRelicSecurity.getAgent().registerOperation(operation);
-
-            return operation;
-        } catch (Throwable e) {
-            if (e instanceof NewRelicSecurityException) {
-                e.printStackTrace();
-                throw e;
-            }
-        }
-        return null;
-    }
-
-    public static void registerExitOperation(AbstractOperation operation) {
-        try {
-            if (operation == null || !NewRelicSecurity.isHookProcessingActive() || NewRelicSecurity.getAgent().getSecurityMetaData().getRequest().isEmpty()
-            ) {
-                return;
-            }
-            NewRelicSecurity.getAgent().registerExitEvent(operation);
-        } catch (Throwable ignored) {
-        }
-    }
-}
