+++ instrumentation-security/low-severity-hooks/src/main/java/com/nr/agent/security/random/javax/crypto/KeyGenerator_Instrumentation.java
@@ -1,29 +1,31 @@
-package java.security;
+package com.nr.agent.security.random.javax.crypto;
 
 import com.newrelic.api.agent.NewRelic;
 import com.newrelic.api.agent.security.NewRelicSecurity;
 import com.newrelic.api.agent.security.schema.AbstractOperation;
 import com.newrelic.api.agent.security.schema.SecurityMetaData;
 import com.newrelic.api.agent.security.schema.StringUtils;
+import com.newrelic.api.agent.security.schema.VulnerabilityCaseType;
 import com.newrelic.api.agent.security.schema.exceptions.NewRelicSecurityException;
 import com.newrelic.api.agent.security.schema.operation.HashCryptoOperation;
 import com.newrelic.api.agent.weaver.MatchType;
 import com.newrelic.api.agent.weaver.Weave;
 import com.newrelic.api.agent.weaver.Weaver;
 
+import java.security.Provider;
+
 import static com.newrelic.api.agent.security.instrumentation.helpers.LowSeverityHelper.DEFAULT;
 import static com.newrelic.api.agent.security.instrumentation.helpers.LowSeverityHelper.LOW_SEVERITY_HOOKS_ENABLED;
 
-@Weave(type = MatchType.ExactClass, originalName = "java.security.MessageDigest")
-public class MessageDigest_Instrumentation {
-
-    public static MessageDigest getInstance(String algorithm) {
+@Weave(type = MatchType.ExactClass, originalName = "javax.crypto.KeyGenerator")
+public class KeyGenerator_Instrumentation {
+    public static final KeyGenerator_Instrumentation getInstance(String algorithm) {
         AbstractOperation operation = null;
         boolean isOwaspHookEnabled = NewRelic.getAgent().getConfig().getValue(LOW_SEVERITY_HOOKS_ENABLED, DEFAULT);
         if (isOwaspHookEnabled){
-            operation = preprocessSecurityHook(algorithm, StringUtils.EMPTY, MessageDigest.class.getName(), "getInstance");
+            operation = preprocessSecurityHook(algorithm, StringUtils.EMPTY, KeyGenerator_Instrumentation.class.getName(), "getInstance", "KEYGENERATOR");
         }
-        MessageDigest returnValue = null;
+        KeyGenerator_Instrumentation returnValue = null;
         try {
             returnValue = Weaver.callOriginal();
         } finally {
@@ -34,13 +36,13 @@ public class MessageDigest_Instrumentation {
         return returnValue;
     }
 
-    public static MessageDigest getInstance(String algorithm, String provider) {
+    public static final KeyGenerator_Instrumentation getInstance(String algorithm, String provider) {
         AbstractOperation operation = null;
         boolean isOwaspHookEnabled = NewRelic.getAgent().getConfig().getValue(LOW_SEVERITY_HOOKS_ENABLED, DEFAULT);
         if (isOwaspHookEnabled){
-            operation = preprocessSecurityHook(algorithm, provider, MessageDigest.class.getName(), "getInstance");
+            operation = preprocessSecurityHook(algorithm, provider, KeyGenerator_Instrumentation.class.getName(), "getInstance", "KEYGENERATOR");
         }
-        MessageDigest returnValue = null;
+        KeyGenerator_Instrumentation returnValue = null;
         try {
             returnValue = Weaver.callOriginal();
         } finally {
@@ -51,13 +53,13 @@ public class MessageDigest_Instrumentation {
         return returnValue;
     }
 
-    public static MessageDigest getInstance(String algorithm, Provider provider) {
+    public static final KeyGenerator_Instrumentation getInstance(String algorithm, Provider provider) {
         AbstractOperation operation = null;
         boolean isOwaspHookEnabled = NewRelic.getAgent().getConfig().getValue(LOW_SEVERITY_HOOKS_ENABLED, DEFAULT);
         if (isOwaspHookEnabled){
-            operation = preprocessSecurityHook(algorithm, provider.getClass().getSimpleName(), MessageDigest.class.getName(), "getInstance");
+            operation = preprocessSecurityHook(algorithm, provider.getClass().getSimpleName(), KeyGenerator_Instrumentation.class.getName(), "getInstance", "KEYGENERATOR");
         }
-        MessageDigest returnValue = null;
+        KeyGenerator_Instrumentation returnValue = null;
         try {
             returnValue = Weaver.callOriginal();
         } finally {
@@ -68,7 +70,7 @@ public class MessageDigest_Instrumentation {
         return returnValue;
     }
 
-    private static AbstractOperation preprocessSecurityHook(String algorithm, String provider, String className, String methodName) {
+    private static AbstractOperation preprocessSecurityHook(String algorithm, String provider, String className, String methodName, String category) {
         try {
             SecurityMetaData securityMetaData = NewRelicSecurity.getAgent().getSecurityMetaData();
             if (!NewRelicSecurity.isHookProcessingActive() || securityMetaData.getRequest().isEmpty()
@@ -76,10 +78,10 @@ public class MessageDigest_Instrumentation {
                 return null;
             }
 
-            HashCryptoOperation operation;
-            operation = new HashCryptoOperation(algorithm, className, methodName);
+            HashCryptoOperation operation = new HashCryptoOperation(algorithm, className, methodName, VulnerabilityCaseType.CRYPTO);
             operation.setProvider(provider);
             operation.setLowSeverityHook(true);
+            operation.setEventCategory(category);
 
             NewRelicSecurity.getAgent().registerOperation(operation);
